requirement.txt
Click==7.0
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.18.1
Pillow==7.0.0
